[QP] Properly escape SQL `LIKE` metachars for `:contains` and friends
Fixes #67075.

`:starts-with`, `:contains` and `:ends-with`, which use SQL `LIKE`
clauses on SQL databases, were previously using the literal string from
the user directly, without escaping the `LIKE` metacharacters `%`, `_`
and backslash.

This change adds a new `sql.qp` multimethod to the driver API, with a
default implementation that uses backspaces and works for most SQL
databases, plus an implementation for SQL Server that uses `[%]`
character classes instead since it doesn't escape with backslashes.

Updates the driver changelog, though fortunately the default case should
work for most drivers. Not to be backported, because of the driver API
change.